requirements.txt
pillow
redis
smbus
RPi.GPIO